requirements.txt
tweepy==4.14.0
pandas==1.5.0
textblob==0.17.1
scikit-learn==1.2.0
python-dotenv==1.0.0
